65일차 비밀번호 암호화, 탈퇴 후 자동로그아웃, 게시물 삭제, 관리자 권한 부여, 검색란, 페이징, 댓글 C

+++ backend/src/main/java/com/example/backend/board/service/BoardService.java
@@ -1,15 +1,19 @@
 package com.example.backend.board.service;
 
-import com.example.backend.board.dto.BoardListInfo;
+import com.example.backend.board.dto.BoardListDto;
 import com.example.backend.board.entity.Board;
 import com.example.backend.board.repository.BoardRepository;
 import com.example.backend.board.dto.BoardDto;
+import com.example.backend.member.entity.Member;
+import com.example.backend.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
-import java.util.Optional;
+import java.util.Map;
 
 @Service
 @RequiredArgsConstructor
@@ -17,15 +21,20 @@ import java.util.Optional;
 public class BoardService {
 
     private final BoardRepository boardRepository;
+    private final MemberRepository memberRepository;
 
-    public void add(BoardDto dto) {
+    public void add(BoardDto dto, Authentication authentication) {
+        if (authentication == null) {
+            throw new RuntimeException("권한이 없습니다.");
+        }
         // entity에 dto의 값들 옮겨 닮고,
         // entity 만들때 테이블이다 생각하고 만들기
         Board board = new Board();
         board.setTitle(dto.getTitle());
         board.setContent(dto.getContent());
-        board.setAuthor(dto.getAuthor());
-
+//        board.setAuthor(authentication.getName());
+        Member author = memberRepository.findById(authentication.getName()).get();
+        board.setAuthor(author);
         // repository에 save 실행
         boardRepository.save(board);
 
@@ -38,42 +47,66 @@ public class BoardService {
         if (dto.getContent() == null || dto.getContent().trim().isBlank()) {
             return false;
         }
-        if (dto.getAuthor() == null || dto.getAuthor().trim().isBlank()) {
-            return false;
-        }
-
         return true;
     }
 
-    public List<BoardListInfo> list() {
-        return boardRepository.findAllByOrderByIdDesc();
+    public Map<String, Object> list(String keyword, Integer pageNumber) {
+//        return boardRepository.findAllByOrderByIdDesc();
+        Page<BoardListDto> boardListDtoPage = boardRepository.findAllBy(keyword, PageRequest.of(pageNumber - 1, 10));
+
+        int totalPages = boardListDtoPage.getTotalPages(); // 마지막 페이지
+        int rightPageNumber = ((pageNumber - 1) / 10 + 1) * 10;
+        int leftPageNumber = rightPageNumber - 9;
+        rightPageNumber = Math.min(rightPageNumber, totalPages);
+        leftPageNumber = Math.max(leftPageNumber, 1);
+
+        var pageInfo = Map.of("totalPages", totalPages,
+                "rightPageNumber", rightPageNumber,
+                "leftPageNumber", leftPageNumber,
+                "currentPageNumber", pageNumber);
+
+        return Map.of("pageInfo", pageInfo, "boardList", boardListDtoPage.getContent());
     }
+
 
     public BoardDto getBoardById(Integer id) {
-        Optional<Board> board = boardRepository.findById(id);
-        BoardDto boardDto = new BoardDto();
-        boardDto.setId(board.get().getId());
-        boardDto.setTitle(board.get().getTitle());
-        boardDto.setContent(board.get().getContent());
-        boardDto.setAuthor(board.get().getAuthor());
-        boardDto.setInsertedAt(board.get().getInsertedAt());
-        return boardDto;
+        BoardDto board = boardRepository.findBoardById(id);
+//        BoardDto boardDto = new BoardDto();
+//        boardDto.setId(board.get().getId());
+//        boardDto.setTitle(board.get().getTitle());
+//        boardDto.setContent(board.get().getContent());
+//        boardDto.setAuthor(board.getAuthor());
+//        boardDto.setInsertedAt(board.get().getInsertedAt());
+        return board;
     }
 
-    public void deleteById(Integer id) {
-        boardRepository.deleteById(id);
+    public void deleteById(Integer id, Authentication authentication) {
+        if (authentication == null) {
+            throw new RuntimeException("권한이 없습니다.");
+        }
+        Board db = boardRepository.findById(id).get();
+        if (db.getAuthor().getEmail().equals(authentication.getName())) {
+            boardRepository.deleteById(id);
+        } else {
+            throw new RuntimeException("권한이 없습니다.");
+        }
+
     }
 
-    public void update(BoardDto boardDto) {
+    public void update(BoardDto boardDto, Authentication authentication) {
+
         // 조회는 조심히, 조회 후 수정 해야함.
         Board db = boardRepository.findById(boardDto.getId()).get();
+        if (db.getAuthor().getEmail().equals(authentication.getName())) {
 
-        // 변경
-        db.setTitle(boardDto.getTitle());
-        db.setContent(boardDto.getContent());
-        db.setAuthor(boardDto.getAuthor());
-
-        // 저장
-        boardRepository.save(db);
+            // 변경
+            db.setTitle(boardDto.getTitle());
+            db.setContent(boardDto.getContent());
+//            db.setAuthor(boardDto.getAuthor());
+            // 저장
+            boardRepository.save(db);
+        } else {
+            throw new RuntimeException("권한이없습니다.");
+        }
     }
 }
